Encapsulate common integration test configuraiton in a trait
